requirements.txt
streamlit>=1.30.0
rembg>=2.0.50
Pillow>=10.0.0
